Conflicts:
	Team8SMS/src/main/java/sg/nus/iss/team8/demo/controllers/FacultyController.java
	Team8SMS/src/main/java/sg/nus/iss/team8/demo/repositories/FacultyRepository.java
	Team8SMS/src/main/java/sg/nus/iss/team8/demo/services/FacultyServiceImplementation.java
	Team8SMS/src/main/resources/application.properties
	Team8SMS/src/main/resources/templates/faculty_grade.html

+++ Team8SMS/src/main/java/sg/nus/iss/team8/demo/controllers/AdminController.java
@@ -1,24 +1,28 @@
 package sg.nus.iss.team8.demo.controllers;
 
-import java.text.DateFormat;
+import java.io.IOException;
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
-import java.time.format.DateTimeFormatter;
+import java.time.YearMonth;
 import java.util.ArrayList;
 import java.util.Date;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
 import javax.validation.Valid;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.propertyeditors.CustomDateEditor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
-import org.springframework.format.annotation.DateTimeFormat;
 import org.springframework.stereotype.Controller;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.WebDataBinder;
@@ -29,55 +33,89 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import sg.nus.iss.team8.demo.models.Courserun;
 import sg.nus.iss.team8.demo.models.CourserunStudent;
 import sg.nus.iss.team8.demo.models.Department;
 import sg.nus.iss.team8.demo.models.Faculty;
 import sg.nus.iss.team8.demo.models.Leave;
-import sg.nus.iss.team8.demo.models.Leave_PK;
 import sg.nus.iss.team8.demo.models.Semester;
 import sg.nus.iss.team8.demo.models.Status;
 import sg.nus.iss.team8.demo.models.Student;
+import sg.nus.iss.team8.demo.models.UserSession;
 import sg.nus.iss.team8.demo.services.AdminService;
 import sg.nus.iss.team8.demo.services.AdminServiceImpl;
-import sg.nus.iss.team8.demo.services.FacultyService;
-import sg.nus.iss.team8.demo.services.FacultyServiceImplementation;
-import sg.nus.iss.team8.demo.services.StudentService;
+import sg.nus.iss.team8.demo.services.GenerateReportService;
+import sg.nus.iss.team8.demo.services.GenerateReportServiceImpl;
 
 
 @Controller
 public class AdminController {
 
 	private AdminService aService;
+	private GenerateReportService grs;
 
-	// injection of faculty service
 	@Autowired
 	public void setAdminService(AdminServiceImpl aService) {
 		this.aService = aService;
 	}
-
-	@InitBinder
-	private void initUserBinder(WebDataBinder binder) {
+	
+	@Autowired
+	public void setGrs(GenerateReportServiceImpl grs) {
+		this.grs = grs;
 	}
+
+
 	
 	@InitBinder
 	protected void initBinder(WebDataBinder binder) {
-//		binder.addValidators(studentValidator);
-		// binder.addValidators(new ProductValidator());
+		binder.addValidators(new FacultyValidator());
 		binder.registerCustomEditor(Date.class, new CustomDateEditor(new SimpleDateFormat("yyyy-MM-dd"), true));
 	}
-
+	@GetMapping("/administrator")
+	public String getAdmin(Model model, HttpServletRequest request) {
+		try{
+			UserSession user=(UserSession)request.getSession(false).getAttribute("user");
+			if(!user.getName().equals("issl")) throw new Exception("invalid user");
+		}catch(Exception e) {
+			return "redirect:/login/faculty";
+		}
+		
+		ArrayList<Semester> allSemesters = aService.findAllSemsters();
+		Semester currentSemester = aService.currentSemester();
+		int courseApplicationCounter = aService.countPendingCourses();
+		int leaveApplicationCounter = aService.countPendingLeaves();
+		
+		model.addAttribute("courseApplicationCounter", courseApplicationCounter);
+		model.addAttribute("leaveApplicationCounter", leaveApplicationCounter);
+		model.addAttribute("allSemesters", allSemesters);
+		model.addAttribute("currentSemester", currentSemester);
+		return "administrator";
+	}
 	@GetMapping("/facultymanagement")
-	public String getFacultyManagement(Model model) {
+	public String getFacultyManagement(Model model,HttpServletRequest request) {
+		try{
+			UserSession user=(UserSession)request.getSession(false).getAttribute("user");
+			if(!user.getName().equals("issl")) throw new Exception("invalid user");
+		}catch(Exception e) {
+			return "redirect:/login/faculty";
+		}
+		
 		ArrayList<Faculty> flist = new ArrayList<Faculty>();
 		flist.addAll(aService.findAllFaculty());
 		model.addAttribute("faculty", flist);
 		return "facultymanagement";
 	}
-
+	@Transactional
 	@GetMapping("/addfaculty")
-	public String addFaculty(Model model) {
+	public String addFaculty(Model model, HttpServletRequest request) {
+		try{
+			UserSession user=(UserSession)request.getSession(false).getAttribute("user");
+			if(!user.getName().equals("issl")) throw new Exception("invalid user");
+		}catch(Exception e) {
+			return "redirect:/login/faculty";
+		}
 		ArrayList<Status> slist = new ArrayList<Status>();
 		ArrayList<Department> dlist = new ArrayList<Department>();
 		Integer newFacultyId = aService.newFacultyId();
@@ -90,17 +128,41 @@ public class AdminController {
 		model.addAttribute("faculty", f);
 		return "facultyform";
 	}
-
+	//@Transactional
 	@PostMapping("/savefaculty")
-	public String saveFaculty(@Valid @ModelAttribute Faculty faculty, BindingResult bindingResult) {
+	public String saveFaculty(@Valid Faculty faculty,BindingResult bindingResult, HttpServletRequest request, Model model) {
+		try{
+			UserSession user=(UserSession)request.getSession(false).getAttribute("user");
+			if(!user.getName().equals("issl")) throw new Exception("invalid user");
+		}catch(Exception e) {
+			return "redirect:/login/faculty";
+		}
 		if (bindingResult.hasErrors())
-			return "facultyform";
+			{System.out.println("has error");
+			ArrayList<Status> slist = new ArrayList<Status>();
+			ArrayList<Department> dlist = new ArrayList<Department>();
+			Integer newFacultyId = aService.newFacultyId();
+			slist.addAll(aService.findAllStatuses());
+			dlist.addAll(aService.findAllDepartments());
+			model.addAttribute("statuses", slist);
+			model.addAttribute("departments", dlist);
+			model.addAttribute("newFacultyId", newFacultyId);
+			model.addAttribute("faculty", faculty);
+			return "facultyform";}
+		System.out.println("no error");
 		aService.saveFaculty(faculty);
 		return "redirect:/facultymanagement";
 	}
 
+	@Transactional
 	@GetMapping("/editfaculty/{id}")
-	public String editFaculty(Model model, @PathVariable("id") Integer id) {
+	public String editFaculty(Model model, @PathVariable("id") Integer id, HttpServletRequest request) {
+		try{
+			UserSession user=(UserSession)request.getSession(false).getAttribute("user");
+			if(!user.getName().equals("issl")) throw new Exception("invalid user");
+		}catch(Exception e) {
+			return "redirect:/login/faculty";
+		}
 		Faculty f = aService.findFacultyById(id);
 		model.addAttribute("faculty", f);
 		ArrayList<Status> slist = new ArrayList<Status>();
@@ -113,32 +175,35 @@ public class AdminController {
 	}
 
 	@GetMapping("/deletefaculty/{id}")
-	public String deleteFaculty(@PathVariable("id") Integer id) {
+	public String deleteFaculty(@PathVariable("id") Integer id, HttpServletRequest request) {
+		try{
+			UserSession user=(UserSession)request.getSession(false).getAttribute("user");
+			if(!user.getName().equals("issl")) throw new Exception("invalid user");
+		}catch(Exception e) {
+			return "redirect:/login/faculty";
+		}
 		Faculty f = aService.findFacultyById(id);
 		aService.deleteFaculty(f);
 		return "redirect:/facultymanagement";
 	}
 	
-	@GetMapping("/administrator")
-	public String getAdmin(Model model) {
-		
-		ArrayList<Semester> allSemesters = aService.findAllSemsters();
-		Semester currentSemester = aService.currentSemester();
-		allSemesters.add(currentSemester);
-		
-		model.addAttribute("allSemesters", allSemesters);
-		model.addAttribute("currentSemester", currentSemester);
-		return "administrator";
-	}
+
 	
 	@PostMapping("/savecurrentsemester")
-	public String saveCurrentSemester(@ModelAttribute Semester sem, BindingResult bindingResult) {
+	public String saveCurrentSemester(@Valid Semester sem, BindingResult bindingResult, RedirectAttributes redirectAttributes) {
+		redirectAttributes.addFlashAttribute("message", "Failed");
+		redirectAttributes.addFlashAttribute("alertClass", "alert-danger");
+		
 		if (bindingResult.hasErrors()) {
 			return "administrator";
 		}
 		aService.saveSemester(sem);
-		int graduationThreshold = 20;
+		int graduationThreshold = 160;
 		aService.applyGraduatedStatus(sem, graduationThreshold);
+		
+		redirectAttributes.addFlashAttribute("message","Success");
+		redirectAttributes.addFlashAttribute("alertClass","alert-success");
+		
 		return "redirect:/administrator";
 	}
 
@@ -148,17 +213,18 @@ public class AdminController {
 
 		int currentPage = page.orElse(1);
 		int pageSize = size.orElse(5);
-
+		
 		Page<Student> studentPage = aService.pageAllStudents(PageRequest.of(currentPage - 1, pageSize));
-
+		System.out.println("before binding studentPage");
 		model.addAttribute("studentPage", studentPage);
+		System.out.println("after binding studentPage");
 		
 		int totalPages = studentPage.getTotalPages();
         if (totalPages > 0) {
             List<Integer> pageNumbers = IntStream.rangeClosed(1, totalPages).boxed().collect(Collectors.toList());
             model.addAttribute("pageNumbers", pageNumbers);
         }
-        
+        System.out.println("before rendering view");
         return "studentmanagement";
 	}
 
@@ -216,7 +282,7 @@ public class AdminController {
 	}
 	
 	@GetMapping("/{acceptOrReject}/{id}/{courseCode}/{semesterid}")
-	public String rejectCourseApplication(@PathVariable(name="acceptOrReject") String acceptOrReject,
+	public String acceptOrRejectCourseApplication(@PathVariable(name="acceptOrReject") String acceptOrReject,
 			@PathVariable(name="id") int id, 
 			@PathVariable(name="courseCode") String courseCode,
 			@PathVariable(name="semesterid") int semesterid) {
@@ -238,7 +304,6 @@ public class AdminController {
 			@RequestParam("size") Optional<Integer> size) {
 		int currentPage = page.orElse(1);
 		int pageSize = size.orElse(5);
-
 
 		Page<Courserun> courserunPage = aService.pageCourserun(PageRequest.of(currentPage - 1, pageSize));
 
@@ -304,8 +369,15 @@ public class AdminController {
 	public String deleteCourserun(Model model, @PathVariable(name = "courseCode") String courseCode,
 			@PathVariable(name = "semesterid") int semesterid) {
 		Courserun course = aService.findCourserun(courseCode, semesterid);
-		aService.removeCourserun(course);
-		return "redirect:/courserunmanagement";
+		// if courserunstudents is not empty, redirect to /viewcourserun/"+courseCode+"/"+semesterid;
+		ArrayList<CourserunStudent> studentsPerCourse = aService.findStudentsByCourseName(course.getCourseName());
+		if(studentsPerCourse.isEmpty()) {
+			aService.removeCourserun(course);
+			return "redirect:/courserunmanagement";
+		} else {
+			return "redirect:/viewcourserun/"+courseCode+"/"+semesterid;
+		}
+		
 	}
 
 	@GetMapping("/viewcourserun/{courseCode}/{semesterid}")
@@ -383,6 +455,79 @@ public class AdminController {
 		Department d = aService.findDepartmentById(id);
 		aService.deleteDepartment(d);
 		return "redirect:departmentmanagement";
+	}
+	
+	@GetMapping("/admin_movementregister")
+	public String movementRegister(Model model, @RequestParam(required=false, name="yearmonth") String ymstring) {
+		YearMonth ym=YearMonth.now();
+		if(ymstring!=null && !ymstring.isEmpty()) {
+			
+			ym=YearMonth.of(Integer.parseInt(ymstring.substring(0,4)),Integer.parseInt(ymstring.substring(5)));
+		}
+		
+		ArrayList<Leave> leaves=aService.findLeavesByYearMonth(ym);
+		ArrayList<YearMonth> yearMonths=aService.findAllYearMonths(ym);
+		ArrayList<String> username=aService.findAllUserName(leaves);
+		HashMap<String,Leave> usernameLeaves=aService.MergeListToMap(leaves,username);
+		//by default we will display the currentMonth leave
+		model.addAttribute("leaves",usernameLeaves);
+		model.addAttribute("yearMonths", yearMonths);
+		model.addAttribute("selectedmonth",ym);
+		return "admin_movementregister";
+	}
+	
+	@RequestMapping("/downloadCSV/classlist")
+	public void downloadCSV(HttpServletRequest request, HttpServletResponse response, 
+			@RequestParam("coursename") String coursename) throws IOException {
+			// create a list of object
+		
+			ArrayList<CourserunStudent> students = aService.findStudentsByCourseName(coursename);
+			String[] headers=new String[]{"Course Name","Student Id", "Student Name", "Degree", "Mobile", "Email", "Grade", "Status"};
+			
+			grs.ExportCSV(request, response, students, headers);
+	}
+	
+	@GetMapping("/viewstudentcourses/{studentid}")
+	public String viewCourserun(Model model,
+			@PathVariable(name = "studentid") int studentid) {
+		Student student = aService.findStudent(studentid);
+		ArrayList<CourserunStudent> coursesPerStudent = aService.findCoursesByStudentId(studentid);
+
+		model.addAttribute("student", student);
+		model.addAttribute("coursesPerStudent", coursesPerStudent);
+
+		return "viewstudentcourses";
+	}
+	
+	@RequestMapping("/downloadCSV/gpalist")
+	public void downloadCsvGpa(HttpServletRequest request, HttpServletResponse response) throws IOException {
+		
+			// create a list of object
+			double gpa;
+			Map<Double, Student> map = new HashMap<>();
+			
+			List<Student> students = aService.findAllStudents();
+			
+			for (Student student : students) {
+				gpa = aService.calculateGPA(student);
+				map.put(gpa, student);
+			}
+			
+			List<String> underGradsWithGpa = map.entrySet()
+												.stream()
+												.map(entry -> entry.getValue().getStudentId() + "," + 
+																entry.getValue().getName() + "," + 
+																entry.getValue().getGender() + "," + 
+																entry.getValue().getMobile() + "," + 
+																entry.getValue().getEmail() + "," + 
+																entry.getValue().getSemester().getLabel() + "," + 
+																entry.getValue().getStatus().getLabel() + "," + 
+																entry.getKey())
+												.sorted()
+												.collect(Collectors.toList());
+			
+			String[] headers=new String[]{"Student_ID","Student_Name", "Gender", "Mobile", "Email", "Semester", "Status", "GPA"};
+			grs.ExportCSV(request, response, underGradsWithGpa, headers);
 	}
 
 }
